feat: update transaction-related methods to use application DTOs (issue #209) (#214)
* feat: update transaction-related methods to use application DTOs (issue #209)

This PR implements issue #209, updating 5 transaction-related methods in the
Bitcoiner interface to use application-layer DTOs instead of infrastructure types.

## Changes

### Updated Methods
1. `GetTransactionByTxID(txID string) (*bitcoindto.TransactionResult, error)`
2. `DecodeRawTransaction(hexTx string) (*bitcoindto.RawTransaction, error)`
3. `FundRawTransaction(hex string) (*bitcoindto.FundRawTransactionResult, error)`
4. `SignRawTransaction(tx *wire.MsgTx, prevtxs []bitcoindto.PreviousTx) (*wire.MsgTx, bool, error)`
5. `SignRawTransactionWithKey(tx *wire.MsgTx, privKeysWIF []string, prevtxs []bitcoindto.PreviousTx) (*wire.MsgTx, bool, error)`

### Implementation
- Added 4 new mapper functions in `btc/mapper.go`:
  - `ToTransactionResult` - Converts GetTransactionResult with proper amount handling
  - `ToRawTransaction` - Converts TxRawResult to RawTransaction DTO
  - `ToFundRawTransactionResult` - Converts funding result with fee conversion
  - `FromPreviousTx` - Converts application DTO to infrastructure type (reverse mapping)

- Updated BTC infrastructure implementation in `transaction.go`:
  - All 5 methods now return DTO types using mappers
  - Proper float64 → btcutil.Amount conversions for currency fields
  - SignRawTransaction and SignRawTransactionWithKey convert PreviousTx DTOs

- Fixed type mismatch in monitor_transaction.go (int64 vs uint64 comparison)

- Regenerated mocks with mockery

### Verification
- Build compiles successfully
- All tests pass
- Linting clean (0 issues)
- No circular dependencies

Part of #203 - Remove infrastructure dependencies from Bitcoiner interface



* refactor: address PR review comments for transaction methods (#214)

This commit addresses three review comments from PR #214:

1. **WalletConflicts type assertion (mapper.go)**
   - Made type assertion more robust by returning an error if the type is not string
   - Prevents silent data corruption by ensuring data integrity
   - Previously silently set empty string on type mismatch

2. **Fee conversion precision fix (mapper.go)**
   - Changed ToFundRawTransactionResult to directly cast int64 to btcutil.Amount
   - Removed unnecessary float64 conversion that risked precision issues
   - Fee is already in satoshis, so float64 conversion was redundant and risky
   - Simplified function signature by removing unnecessary error return

3. **Refactor duplicate code in SignRawTransaction methods (transaction.go)**
   - Created helper functions to reduce code duplication:
     - prepareSignRawTransactionInputs: Prepares common inputs (hex tx, marshaled prevtxs)
     - processSignRawTransactionResult: Processes sign result with error handling
   - Both SignRawTransaction and SignRawTransactionWithKey now use these helpers
   - Removed 40+ lines of duplicated code
   - Improved maintainability and consistency

All changes maintain backward compatibility and pass verification:
- Build: Success
- Linting: 0 issues
- Tests: All passing



---------

+++ internal/application/usecase/watch/btc/monitor_transaction.go
@@ -187,7 +187,7 @@ func (u *monitorTransactionUseCase) checkTransactionConfirmation(
 		"required", u.btcClient.ConfirmationBlock())
 
 	// Check if confirmations meet threshold
-	if tx.Confirmations >= u.btcClient.ConfirmationBlock() {
+	if uint64(tx.Confirmations) >= u.btcClient.ConfirmationBlock() {
 		return true, nil
 	}
